requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.0
django-mysql==2.2.2
djangorestframework==3.7.7
idna==2.8
more-itertools==6.0.0
oauthlib==3.0.1
Pillow==5.4.1
pluggy==0.8.1
py==1.7.0
PyJWT==1.5.3
pytest==4.2.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
urllib3==1.24.1
